Experimental: Unify NamingSchemes.
Naming schemes only differ on three parameters:
 1. characters allowed at beginning of name
 2. characters allowed elsewhere
 3. minimum length of name

All other considerations are determined by the requirements of python
syntax and readability. Re-work the naming schemes to reflect this.

Many odd discrepencies among the naming schemes have been rectified. For
the most part these were bugs, but may have been due to legacy inertia
to some extend. This change will bother users by calling out some of
their bizarre naming conventions, but all such issues are valid.

One major improvement not seen here would be to set a "minimum length"
for names, and check it and flag it separately from formatting issues.
This will greatly reduce the need for users to read and understand
regex, will also be much more actionable, but will also remove this
additional resonsibility from the naming schemes. Most of the weirdness
that remains is due to minimum-length requirements, and separating that
concern will make naming schemes quite simple.

For churn considerations, I've not modified the names of the naming
scheme objects although they've transitioned from being classes to
objects, and there's a couple other bits of odd naming in here for the
same reason. My aim was chiefly to show that this was possible, and spur
consideration of the idea.